codesearch(obs): C2 fix — http.end targetHash and shared indexingLog helper
